Add secondary log scanning to verify gpinitsystem successfullness.
Scan through the gpinitsystem logfile and verify the installation was
successful.  Do this by verifying no messages were logged with a error
level above info.  If there are messages with a warning or higher, failure
consensus is reached and the installation is marked as failed.  If there
aren't, the installation is considered successful in spite of the non-zero
exit code given by gpinitsystem.