Fix execute calls to work with Ambari 2.0.
Start specifying a correct path in the Execute call.  Don't rely on the one
given by the sourcing greenplum_path.sh in .bashrc.  ~/.bashrc and
greenplum_path.sh are both called/sourced, but for whatever reason even
with the path being updated the relevant commands can not be found.

This is break between Ambari 1.7.x and 2.0.x.